requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyflakes>=3.0.0
